# Switch to use Erlang nifs in frame mask application.