I'd like to throw this strawman up to see if we can maybe switch the XOR work to be done using built-in Erlang nifs.

I _think_ this should result in fewer allocations and should use native code as much as possible for the bulk XORing.

Let me know what y'all think. :)