@ryanwinchester 2023-12-07
```elixir
defmodule Original do
  # Masking is done @mask_size bits at a time until there is less than that number of bits left.
  # We then go 32 bits at a time until there is less than 32 bits left. We then go 8 bits at
  # a time. This yields some significant perforamnce gains for only marginally more complexity
  @mask_size 512

  # Note that masking is an involution, so we don't need a separate unmask function
  def mask(payload, mask) when bit_size(payload) >= @mask_size do
    payload
    |> do_mask(String.duplicate(<<mask::32>>, div(@mask_size, 32)), [])
    |> Enum.reverse()
    |> IO.iodata_to_binary()
  end

  def mask(payload, mask) do
    payload
    |> do_mask(<<mask::32>>, [])
    |> Enum.reverse()
    |> IO.iodata_to_binary()
  end

  defp do_mask(
         <<h::unquote(@mask_size), rest::binary>>,
         <<int_mask::unquote(@mask_size)>> = mask,
         acc
       ) do
    do_mask(rest, mask, [<<Bitwise.bxor(h, int_mask)::unquote(@mask_size)>> | acc])
  end

  defp do_mask(<<h::32, rest::binary>>, <<int_mask::32, _mask_rest::binary>> = mask, acc) do
    do_mask(rest, mask, [<<Bitwise.bxor(h, int_mask)::32>> | acc])
  end

  defp do_mask(<<h::8, rest::binary>>, <<current::8, mask::24, _mask_rest::binary>>, acc) do
    do_mask(rest, <<mask::24, current::8>>, [<<Bitwise.bxor(h, current)::8>> | acc])
  end

  defp do_mask(<<>>, _mask, acc), do: acc
end
```
```elixir
defmodule New do
  # note that we assume the mask integer is 32-bits and do not check
  @spec mask(binary(), integer()) :: binary()
  def mask(payload, mask_integer) do

    # 1. Allocate the binary for the mask repetitions
    payload_size = byte_size(payload)
                
    mask_repetitions = case { div(payload_size,4), rem(payload_size,4)} do
      {count_4_bytes, 0 } ->
        count_4_bytes # payload length is a multiple of 4, so the mask will be fine
      {count_4_bytes, _count_stragglers} ->
       count_4_bytes + 1 # bump up by one mask size
    end
        
    mask_binary = :binary.copy( <<mask_integer::32>>, mask_repetitions)
              
    # 2. Trim the binary if needed
    fit_mask = :binary.part(mask_binary, 0, payload_size)
        
    # 3. XOR (in a nif) the payload and mask binary
    masked_payload = :crypto.exor(payload, fit_mask)
  end
end
```
```elixir
input_sm = String.duplicate("a", 10_000)
input_md = String.duplicate("a", 200_000)
input_lg = String.duplicate("a", 10_000_000)

Benchee.run(
  %{
    "original" => fn input -> Original.mask(input, 1234) end,
    "proposed" => fn input -> New.mask(input, 1234) end
  },
  inputs: %{
    "sm" => input_sm,
    "md" => input_md,
    "lg" => input_lg,
  },
  time: 5,
  memory_time: 2
)
```
```
$ mix run bench/mask.exs

Operating System: macOS
CPU Information: Apple M2 Ultra
Number of Available Cores: 24
Available memory: 128 GB
Elixir 1.15.7
Erlang 26.1.2

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
memory time: 2 s
reduction time: 0 ns
parallel: 1
inputs: lg, md, sm
Estimated total run time: 54 s

Benchmarking original with input lg ...
Benchmarking original with input md ...
Benchmarking original with input sm ...
Benchmarking proposed with input lg ...
Benchmarking proposed with input md ...
Benchmarking proposed with input sm ...

##### With input lg #####
Name               ips        average  deviation         median         99th %
proposed         62.47       16.01 ms     ±2.43%       15.98 ms       16.92 ms
original         19.96       50.10 ms     ±8.09%       49.28 ms       58.74 ms

Comparison:
proposed         62.47
original         19.96 - 3.13x slower +34.09 ms

Memory usage statistics:

Name        Memory usage
proposed        0.152 MB
original        43.89 MB - 287.84x memory usage +43.74 MB

**All measurements for memory usage were the same**

##### With input md #####
Name               ips        average  deviation         median         99th %
proposed        3.20 K      312.60 μs     ±5.73%      307.96 μs      373.50 μs
original        1.73 K      576.40 μs    ±11.81%      574.40 μs      767.73 μs

Comparison:
proposed        3.20 K
original        1.73 K - 1.84x slower +263.80 μs

Memory usage statistics:

Name        Memory usage
proposed         2.88 KB
original       857.89 KB - 297.59x memory usage +855.01 KB

**All measurements for memory usage were the same**

##### With input sm #####
Name               ips        average  deviation         median         99th %
proposed       63.83 K       15.67 μs    ±26.38%          15 μs       30.88 μs
original       29.06 K       34.42 μs   ±193.00%       20.38 μs      395.46 μs

Comparison:
proposed       63.83 K
original       29.06 K - 2.20x slower +18.75 μs

Memory usage statistics:

Name        Memory usage
proposed        0.117 KB
original        45.66 KB - 389.60x memory usage +45.54 KB

**All measurements for memory usage were the same**
```