@ryanwinchester 2023-12-08
Attempt No. 2

```elixir
input = String.duplicate("a", 1000)

Benchee.run(
  %{
    "original" => fn ->
      for _ <- 1..100_000, do: Original.mask(input, 1234)
    end,
    "proposed" => fn ->
      for _ <- 1..100_000, do: New.mask(input, 1234)
    end
  },
  time: 5,
  memory_time: 2
)
```

```
$ mix run bench/mask.exs
warning: variable "masked_payload" is unused (if the variable is not meant to be used, prefix it with an underscore)
  bench/mask.exs:62: New.mask/2

Operating System: macOS
CPU Information: Apple M2 Ultra
Number of Available Cores: 24
Available memory: 128 GB
Elixir 1.15.7
Erlang 26.1.2

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
memory time: 2 s
reduction time: 0 ns
parallel: 1
inputs: none specified
Estimated total run time: 18 s

Benchmarking original ...
Benchmarking proposed ...

Name               ips        average  deviation         median         99th %
proposed          4.80      208.30 ms     ±3.07%      207.93 ms      221.18 ms
original          2.10      475.64 ms     ±1.37%      477.59 ms      483.00 ms

Comparison:
proposed          4.80
original          2.10 - 2.28x slower +267.34 ms

Memory usage statistics:

Name        Memory usage
proposed        12.97 MB
original       534.82 MB - 41.24x memory usage +521.85 MB

**All measurements for memory usage were the same**
```