@mtrudel 2023-12-08
I find the results hard to believe (less memory usage even though we're doing one giant allocation?), but IMO it's a huge win! I wonder where else we'd benefit from this approach.

If we can get this cleaned up & it bears fruit in benchmark CI I'd be happy to merge it!