@ryanwinchester 2023-12-08
> I find the results hard to believe (less memory usage even though we're doing one giant allocation?)

Yes, investigating the error, function name changes messed me up and my benchmarks are a giant flub. Fixing and redoing them.